Fixed a bug in VolumeContains(used uint instead of unsigned long to compare hashes). HNode Coarsening seems to work nicely in 2d. In 3d the selection is not yet perfectly adjusted. Fixed compile-problems in TKDPlugin. Adjusted TKDPlugin CMakeLists.txt to only optionally build TKDTestsuite. Added static int dim to VertexBase, EdgeBase, Face and Volume

+++ ugbase/lib_grid/grid/grid_util.cpp
@@ -617,7 +617,7 @@ bool VolumeContains(Volume* v, FaceVertices* fv)
 {
 	FaceDescriptor fd;
 	uint numFaces = v->num_faces();
-	uint hash = hash_key(fv);
+	unsigned long hash = hash_key(fv);
 	for(uint i = 0; i < numFaces; ++i)
 	{
 		v->face(i, fd);
